# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestionar kits de sensores y sus datos.",
        "title": "API Hexagonal Go (Sensor Kits)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.tu-soporte.com/support",
            "email": "support@tu-dominio.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080  // Cambia esto por tu host y puerto reales (puedes usar variables de entorno)",
    "basePath": "/v1             // Base path de tu API v1",
    "paths": {
...
        "/data/light/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new light level data record for a specific kit.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Light Data"
                ],
                "summary": "Register light data record",
                "parameters": [
                    {
                        "description": "Light data record",
                        "name": "lightData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RegisterLightRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Light record registered successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entities.LightData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body, validation failed, or invalid Kit ID",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (token missing or invalid)",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error while registering light record",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
...
        "/data/motion/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new motion level data record for a specific kit.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motion Data"
                ],
                "summary": "Register motion data record",
                "parameters": [
                    {
                        "description": "motion data record",
                        "name": "motionData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RegisterMotionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "motion record registered successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entities.MotionData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body, validation failed, or invalid Kit ID",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (token missing or invalid)",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error while registering motion record",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/data/motion/kit/{kit_id}/minutes/{minutes}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves motion data records for a specific kit within the last N minutes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Motion Data"
                ],
                "summary": "Get recent motion data records",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Kit ID",
                        "name": "kit_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Number of minutes ago to fetch records from",
                        "name": "minutes",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "motion records retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entities.MotionData"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Kit ID or minutes parameter provided",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (token missing or invalid)",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error while retrieving motion records",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.LightData": {
            "type": "object",
            "properties": {
                "data_id": {
                    "description": "Corresponds to data_id PK",
                    "type": "integer"
                },
                "kit_id": {
                    "description": "Foreign key to kits table",
                    "type": "integer"
                },
                "light_level": {
                    "description": "Light level reading (e.g., lux)",
                    "type": "number"
                },
                "timestamp": {
                    "description": "Timestamp from DB default",
                    "type": "string"
                }
            }
        },
        "entities.MotionData": {
            "type": "object",
            "properties": {
                "data_id": {
                    "description": "Corresponds to data_id PK",
                    "type": "integer"
                },
                "kit_id": {
                    "description": "Foreign key to kits table",
                    "type": "integer"
                },
                "motion_detected": {
                    "description": "Whether motion was detected",
                    "type": "boolean"
                },
                "timestamp": {
                    "description": "Timestamp from DB default",
                    "type": "string"
                }
            }
        },
...
        "request.RegisterLightRequest": {
            "type": "object",
            "required": [
                "kit_id",
                "light_level"
            ],
            "properties": {
                "kit_id": {
                    "description": "Kit ID must be provided and positive",
                    "type": "integer"
                },
                "light_level": {
                    "description": "Light level must be provided and non-negative (adjust if needed)",
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "request.RegisterMotionRequest": {
            "type": "object",
            "required": [
                "kit_id"
            ],
            "properties": {
                "kit_id": {
                    "type": "integer"
                },
                "motion_detected": {
                    "description": "No specific validate tag needed, presence checked by ShouldBindJSON",
                    "type": "boolean"
                }
            }
        },
...
        "responses.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}